server/requirements.txt
# FastFit AI Server Dependencies
# ==============================

# ⚠️ PyTorch with CUDA - 반드시 별도 설치 필요!
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu130
# 또는 cu121, cu124 등 본인 CUDA 버전에 맞게 선택

# FastAPI Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Image Processing
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# ML Libraries (torch는 위 명령어로 별도 설치)
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0

# FastFit Core Dependencies
huggingface-hub>=0.20.0
einops>=0.7.0
omegaconf>=2.3.0
onnxruntime>=1.16.0

# FastFit Additional Dependencies (from official repo)
scipy>=1.11.0
fvcore>=0.1.5
cloudpickle>=3.0.0
pycocotools>=2.0.7
av>=12.0.0
dists-pytorch>=0.1
matplotlib>=3.8.0
tqdm>=4.66.0

# Utilities
python-dotenv>=1.0.0